+++ backend/sftp/options.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"path"
+	"path/filepath"
 	"runtime"
 	"strconv"
 
@@ -231,7 +231,10 @@ func findHomeSystemKnownHosts(knownHostsFiles []string) ([]string, error) {
 	if err != nil {
 		return nil, err
 	}
-	homeKnownHostsPath := utils.EnsureLeadingSlash(path.Join(home, ".ssh/known_hosts"))
+	homeKnownHostsPath := filepath.Join(home, ".ssh", "known_hosts")
+	if runtime.GOOS != "windows" {
+		homeKnownHostsPath = utils.EnsureLeadingSlash(homeKnownHostsPath)
+	}
 
 	// check file existence first to prevent auto-vivification of file
 	found, err := foundFile(homeKnownHostsPath)