+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Windows support in the os backend.
+### Fixed
+- Ability to run all unit tests on Windows.
 
 ## [6.24.0] - 2024-12-16
 ### Security